+++ R/checker.R
@@ -5,7 +5,7 @@
 #' @returns Either stops execution with an error message if checks fail, or
 #'          returns NULL invisibly if all checks pass.
 #' @noRd
-checker <- function(...) {
+checker <- function(called_from = NULL, ...) {
   # Capture arguments
   args <- list(...)
   # Extract specific arguments (with default NULL if not provided)
@@ -13,53 +13,87 @@ checker <- function(...) {
   mat <- args$mat
   ns <- args$ns
   n <- args$n
+  penalty <- args$penalty
+  vary <- args$vary
 
   # This section deals with checking data and mat inputs
   if (all(c("data", "mat") %in% names(args))) {
-    if (is.null(data) && is.null(mat)) {
+    if (is.null(data) & is.null(mat)) {
       stop("Either 'data' or 'mat' must be provided.")
-      } else if (!is.null(data) && !is.null(mat)) {
-    message("Both 'data' and 'mat' provided. 'data' will be used.")
-        if (!is.data.frame(data) & !is.matrix(data)) {
-          stop("'data' must be a data frame or matrix.")
-        }
-        if (!all(sapply(data, is.numeric))){
-          stop("All variables in 'data' must be numeric.")
-        }
-      } else if (!is.null(data)) {
-        if (!all(sapply(data, is.numeric))){
-          stop("All variables in 'data' must be numeric.
-               If you have an ordered categorical variable stored as a factor, please convert it to numeric first.
-               This requirement helps prevent accidentally including unordered categorical variables in the calculation.")
-        }
-      } else if (!is.null(mat)) {
-        if (!is.matrix(mat)) {
-          stop("'mat' must be a matrix.")
-        } else if(!isSymmetric(mat)) {
-          stop("'mat' must be a symmetric matrix.")
-        }
-        if (!is.numeric(mat)) {
-          stop("All entries in 'mat' must be numeric")
-        }
-        if (any(is.na(mat))) {
-          stop("'mat' must not contain missing values.")
+    }
+    if (!is.null(data)) {
+      if (!is.data.frame(data) & !is.matrix(data)) {
+        stop("'data' must be a data frame or matrix.")
+      }
+      if (!all(sapply(data, is.numeric))) {
+        stop("All variables in 'data' must be numeric.
+           If you have an ordered categorical variable stored as a factor,
+           please convert it to numeric first. This requirement helps prevent
+           accidentally including unordered categorical variables in the calculation.")
+      }
+    }
+    if (!is.null(mat)) {
+      if (!is.matrix(mat)) {
+        stop("'mat' must be a matrix.")
+      }
+      if (!isSymmetric(mat)) {
+        stop("'mat' must be a symmetric matrix.")
+      }
+      if (!is.numeric(mat)) {
+        stop("All entries in 'mat' must be numeric.")
+      }
+      if (any(is.na(mat))) {
+        stop("'mat' must not contain missing values.")
+      }
+      if (!is.null(rownames(mat)) & !is.null(colnames(mat))) {
+        if (!identical(rownames(mat), colnames(mat))) {
+          stop("Row and column names of 'mat' must be identical.")
         }
       }
+    }
+    if (!is.null(data) & !is.null(mat)) {
+      if (ncol(data) != ncol(mat)) {
+        stop("The number of columns used for network analysis in 'data' must match the number of columns in 'mat'.")
+      }
+      if (!is.null(colnames(data)) & !is.null(colnames(mat))) {
+        if (!identical(colnames(data), colnames(mat))) {
+          stop("Column names of 'data' and 'mat' must match and be in the same order.")
+        }
+      }
+    }
   }
 
-  # This section deals with checking ns for multiple regression model
-  if ("ns" %in% names(args)) {
-    if ("mat" %in% names(args)){
-      if (length(ns) != ncol(mat) & length(ns) != 1) {
-        stop("'ns' must be either a single value or a vector with one entry per column in 'mat'.")
+  if (!is.null(called_from)){
+    if(called_from == "neighborhood"){
+    if ("ns" %in% names(args)){
+      if ("mat" %in% names(args)){
+        if (length(ns) != ncol(mat) & length(ns) != 1){
+          stop("'ns' must be either a single value or a vector with one entry per column in 'mat'.")
+        }
       }
-    }
-    if ("data" %in% names(args)){
-      if (length(ns) != ncol(data) & length(ns) != 1) {
-        stop("'ns' must be either a single value or a vector with one entry per column in 'data'.")
+      if ("data" %in% names(args)){
+        if (length(ns) != ncol(data) & length(ns) != 1){
+          stop("'ns' must be either a single value or a vector with one entry per column in 'data'.")
+        }
       }
     }
   }
+
+  if (called_from == "regularization"){
+  # This section deals with checking ns for multiple regression model
+  if ("ns" %in% names(args)) {
+    if ("mat" %in% names(args)) {
+      if (!(length(ns) == 1 || (is.matrix(ns) && nrow(ns) == ncol(mat) && ncol(ns) == ncol(mat)))) {
+        stop("'ns' must be either a single value or a matrix with dimensions matching 'mat'.")
+      }
+    }
+    if ("data" %in% names(args)) {
+      if (!(length(ns) == 1 || (is.matrix(ns) && nrow(ns) == ncol(data) && ncol(ns) == ncol(data)))) {
+        stop("'ns' must be either a single value or a matrix with dimensions matching the number of columns in 'data'.")
+      }
+    }
+  }
+  }}
 
   # This section deals with checking n for a single regression model
   if ("n" %in% names(args)) {
@@ -67,6 +101,13 @@ checker <- function(...) {
       stop("Length of 'n' must be 1.")
     } else if (!is.numeric(n) || n <= 0) {
       stop("'n' must be a positive numeric value.")
+    }
+  }
+
+  # Check lambda to be varying if the penalty is glasso
+  if (all(c("penalty", "vary") %in% names(args))) {
+    if (penalty == "glasso" & vary != "lambda") {
+      stop("For 'glasso' penalty, 'vary' must be set to 'lambda' as this is the only penalty parameter. If you want to provide your own lambda values you can do this in the corresponding argument 'lambda' but you still have to set 'vary' to 'lambda'.")
     }
   }
 }